rename bridge_main_averaging
mv  bridge_main_averaging to lib. bridge_main
